Machine Learning/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
Pillow==12.0.0
ultralytics==8.3.239
